requirements.txt
pyspark
pandas
numpy
boto3
botocore
psycopg2-binary
pyyaml
